[btree_walker] Revisit shared nodes with summaries
The BTreeWalker avoids reading shared nodes and their descendants
more than once. To inform the NodeVisitor about the shared node
information, the BTreeWalker now caches the summaries for every
visited node so that it could update the NodeVisitor on visiting
shared nodes. The NodeVisitor is responsible for summarizing leaf
nodes, and the BTreeWalker aggregates that information for internal
nodes during traversal.

This way allows the NodeVisitor to summarize a btree even with
shared nodes, like counting the number of mapped blocks in thin_check.