[btree_walker] Defer the visiting of shared nodes while race happens
The calling thread of BTreeWalker::walk() might encounter race if the
visiting node is being checked by another thread. To solve this issue
and avoid performance impact, the BTreeWalker skips those nodes and
returns the block numbers to the caller. The caller then is required to
visit those deferred nodes to complete visiting.

Although there's no significant performance impact, this way exposes the
potential flaws in BTreeWalker and increases the complexity of the
caller.