[thin_check] Make thin_check functionable and add further optimization
- Count the number of mapped blocks of each device
- Fix ref counting on mapping trees
- Revise the data structures to reduce memory consumption
- Remove underfull and key range checking in the first pass
- Check the parent key range in the second pass

The underfull property is path-related, thus it should be checked in
the second stage. Given an underfull node which is errnously referred
as a root and non-root in different trees, the first case should
succeed, and the second case should fail.

The parent key range should be checked in the second pass since we're
comparing the key range against the actual keys gathered from the the
leaves, which are read in the second pass.