[cache_check] Do not perform boundary check on cache mappings
Historically, dm-cache does not invalidate the mappings beyond the
origin's size while the cache origin get reduced, leading to
inconsistencies between the mapping addresses and the size of the
discard bitset. Additionally, the on-disk discard bitset is not resized
immediately during a resume operation; therefore, the inconsistency is
not detected immediately after reducing the origin; it only becomes
apparent during the first suspend on the resized cache, which triggers
sync_metadata().

To maintain backward compatibility, remove the boundary checks on cache
mappings, even though this kernel issue might be fixed in the future.