[btree_walker] Use two workers for reading leaf nodes
The number of leaf workers doesn't greatly affect the efficiency with
AsyncIoEngine, as there are just a few nodes with space map trees
(bitmap tree and ref count trees). However, that could affect
SyncIoEngine's outcome. Here we choose two threads to balance the
performance for SyncIoEngine.

In the release version, we might adjust the thread number at runtime,
according to the engine type.